+++ packages/ui/button/src/Button.stories.tsx
@@ -9,6 +9,7 @@ export default {
     prefix: { control: { type: 'text' } },
     suffix: { control: { type: 'text' } },
     children: { control: { type: 'text' } },
+    spinner: { control: { type: 'text' } },
     disabled: { control: { type: 'boolean' } }
   }
 }
@@ -18,9 +19,12 @@ export const Default = (args: ButtonProps) => (
     {...args}
     classNames={
       args.classNames ?? {
-        root: 'bg-neutral-300 rounded-md border-1 border-neutral-400/50 text-neutral-700 font-medium hover:bg-neutral-400 active:scale-97'
+        btn: 'w-25 h-10 bg-neutral-300 rounded-md border-1 border-neutral-400/50 text-neutral-700 font-medium hover:bg-neutral-400 active:scale-97'
       }
     }
+    spinnerProps={{
+      className: 'w-4.5 h-4.5 border-4 border-solid border-neutral-500 border-t-blue-500'
+    }}
   >
     {args.children ?? 'Button'}
   </Button>